refactor(calculator): 입력 검증 클래스 분리
